Add files via upload
Task 1: Linear Regression - Sparks Foundation

Description:
In this project for The Sparks Foundation, I implemented a Linear Regression model to predict students' scores based on the number of hours they study. The goal was to analyze the relationship between study hours and academic performance. Leveraging Python and its data science libraries, including pandas for data manipulation, numpy for numerical operations, scikit-learn for building the regression model, and Matplotlib along with Seaborn for data visualization, the project was developed within a Jupyter Notebook. This task serves as a practical demonstration of applying linear regression in a real-world scenario.

Technical Stack:

Programming Language: Python
Libraries: pandas, numpy, scikit-learn
Development Environment: Jupyter Notebook
Data Visualization: Matplotlib, Seaborn
This project not only addresses the linear regression problem but also provides an educational resource for understanding the implementation of regression models and their application in predicting outcomes based on input features. The code and analysis can be found in the GitHub repository for further reference and learning.